CASS-1828 Consolidate PriamInstance fetching into a single interface.
