+++ priam/src/main/java/com/netflix/priam/identity/InstanceIdentity.java
@@ -16,7 +16,6 @@
  */
 package com.netflix.priam.identity;
 
-import com.google.common.base.Preconditions;
 import com.google.common.base.Predicate;
 import com.google.common.collect.Iterables;
 import com.google.common.collect.ListMultimap;
@@ -26,22 +25,10 @@ import com.google.inject.Inject;
 import com.google.inject.Singleton;
 import com.netflix.priam.config.IConfiguration;
 import com.netflix.priam.identity.config.InstanceInfo;
-import com.netflix.priam.identity.token.IDeadTokenRetriever;
-import com.netflix.priam.identity.token.INewTokenRetriever;
-import com.netflix.priam.identity.token.IPreGeneratedTokenRetriever;
-import com.netflix.priam.identity.token.TokenRetrieverUtils;
-import com.netflix.priam.identity.token.TokenRetrieverUtils.GossipParseException;
-import com.netflix.priam.utils.ITokenManager;
-import com.netflix.priam.utils.RetryableCallable;
-import com.netflix.priam.utils.Sleeper;
-import java.net.UnknownHostException;
-import java.util.Collections;
+import com.netflix.priam.identity.token.ITokenRetriever;
 import java.util.HashMap;
 import java.util.LinkedList;
 import java.util.List;
-import java.util.Optional;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
 /**
  * This class provides the central place to create and consume the identity of the instance - token,
@@ -49,7 +36,6 @@ import org.slf4j.LoggerFactory;
  */
 @Singleton
 public class InstanceIdentity {
-    private static final Logger logger = LoggerFactory.getLogger(InstanceIdentity.class);
     public static final String DUMMY_INSTANCE_ID = "new_slot";
 
     private final ListMultimap<String, PriamInstance> locMap =
@@ -57,8 +43,6 @@ public class InstanceIdentity {
     private final IPriamInstanceFactory<PriamInstance> factory;
     private final IMembership membership;
     private final IConfiguration config;
-    private final Sleeper sleeper;
-    private final ITokenManager tokenManager;
 
     private final Predicate<PriamInstance> differentHostPredicate =
             new Predicate<PriamInstance>() {
@@ -88,15 +72,9 @@ public class InstanceIdentity {
     private PriamInstance myInstance;
     // Instance information contains other information like ASG/vpc-id etc.
     private InstanceInfo myInstanceInfo;
-    private boolean isReplace = false;
-    private boolean isTokenPregenerated = false;
-    private String replacedIp = "";
-    private final IDeadTokenRetriever deadTokenRetriever;
-    private final IPreGeneratedTokenRetriever preGeneratedTokenRetriever;
-    private final INewTokenRetriever newTokenRetriever;
-
-    private final java.util.function.Predicate<PriamInstance> sameHostPredicate =
-            (i) -> i.getInstanceId().equals(myInstanceInfo.getInstanceId());
+    private boolean isReplace;
+    private boolean isTokenPregenerated;
+    private String replacedIp;
 
     @Inject
     // Note: do not parameterized the generic type variable to an implementation as
@@ -106,23 +84,17 @@ public class InstanceIdentity {
             IPriamInstanceFactory factory,
             IMembership membership,
             IConfiguration config,
-            Sleeper sleeper,
-            ITokenManager tokenManager,
-            IDeadTokenRetriever deadTokenRetriever,
-            IPreGeneratedTokenRetriever preGeneratedTokenRetriever,
-            INewTokenRetriever newTokenRetriever,
-            InstanceInfo instanceInfo)
+            InstanceInfo instanceInfo,
+            ITokenRetriever tokenRetriever)
             throws Exception {
         this.factory = factory;
         this.membership = membership;
         this.config = config;
-        this.sleeper = sleeper;
-        this.tokenManager = tokenManager;
-        this.deadTokenRetriever = deadTokenRetriever;
-        this.preGeneratedTokenRetriever = preGeneratedTokenRetriever;
-        this.newTokenRetriever = newTokenRetriever;
         this.myInstanceInfo = instanceInfo;
-        init();
+        this.myInstance = tokenRetriever.get();
+        this.replacedIp = tokenRetriever.getReplacedIp();
+        this.isReplace = tokenRetriever.isReplace();
+        this.isTokenPregenerated = tokenRetriever.isTokenPregenerated();
     }
 
     public PriamInstance getInstance() {
@@ -133,185 +105,6 @@ public class InstanceIdentity {
         return myInstanceInfo;
     }
 
-    public void init() throws Exception {
-        // Grab the token which was preassigned.
-        logger.info("trying to grab preassigned token.");
-        myInstance = grabPreAssignedToken();
-
-        // Grab a dead token.
-        if (myInstance == null) {
-            logger.info("unable to grab preassigned token. trying to grab a dead token.");
-            myInstance = grabDeadToken();
-        }
-
-        // Grab a pre-generated token if there is such one.
-        if (myInstance == null) {
-            logger.info("unable to grab a dead token. trying to grab a pregenerated token.");
-            myInstance = grabPreGeneratedToken();
-        }
-
-        // Grab a new token
-        if (myInstance == null) {
-            logger.info("unable to grab a pregenerated token. trying to grab a new token.");
-            myInstance = grabNewToken();
-        }
-
-        logger.info("My token: {}", myInstance.getToken());
-    }
-
-    private PriamInstance grabPreAssignedToken() throws Exception {
-        return new RetryableCallable<PriamInstance>() {
-            @Override
-            public PriamInstance retriableCall() throws Exception {
-                // Check if this node is decommissioned.
-                List<PriamInstance> deadInstances =
-                        factory.getAllIds(config.getAppName() + "-dead");
-                PriamInstance instance =
-                        findInstance(deadInstances, sameHostPredicate).orElse(null);
-                if (instance != null) {
-                    instance.setOutOfService(true);
-                }
-
-                if (instance == null) {
-                    List<PriamInstance> aliveInstances = factory.getAllIds(config.getAppName());
-                    instance = findInstance(aliveInstances, sameHostPredicate).orElse(null);
-
-                    if (instance != null) {
-                        instance.setOutOfService(false);
-
-                        // Priam might have crashed before bootstrapping Cassandra in replace mode.
-                        // So, it is premature to use the assigned token without checking Cassandra
-                        // gossip.
-
-                        // Infer current ownership information from other instances using gossip.
-                        TokenRetrieverUtils.InferredTokenOwnership inferredTokenOwnership =
-                                TokenRetrieverUtils.inferTokenOwnerFromGossip(
-                                        aliveInstances, instance.getToken(), instance.getDC());
-                        // if unreachable rely on token database.
-                        // if mismatch rely on token database.
-                        if (inferredTokenOwnership.getTokenInformationStatus()
-                                == TokenRetrieverUtils.InferredTokenOwnership.TokenInformationStatus
-                                        .GOOD) {
-                            Preconditions.checkNotNull(
-                                    inferredTokenOwnership.getTokenInformation());
-                            String inferredIp =
-                                    inferredTokenOwnership.getTokenInformation().getIpAddress();
-                            if (!inferredIp.equalsIgnoreCase(instance.getHostIP())) {
-                                if (inferredTokenOwnership.getTokenInformation().isLive()) {
-                                    throw new GossipParseException(
-                                            "We have been assigned a token that C* thinks is alive. Throwing to buy time in the hopes that Gossip just needs to settle.");
-                                }
-                                setReplacedIp(inferredIp);
-                                logger.info(
-                                        "Priam found that the token is not alive according to Cassandra and we should start Cassandra in replace mode with replace ip: "
-                                                + inferredIp);
-                            }
-                        }
-                    }
-                }
-
-                if (instance != null) {
-                    logger.info(
-                            "{} found that this node is {}."
-                                    + " application: {},"
-                                    + " id: {},"
-                                    + " instance: {},"
-                                    + " region: {},"
-                                    + " host ip: {},"
-                                    + " host name: {},"
-                                    + " token: {}",
-                            instance.isOutOfService() ? "[Dead]" : "[Alive]",
-                            instance.isOutOfService() ? "dead" : "alive",
-                            instance.getApp(),
-                            instance.getId(),
-                            instance.getInstanceId(),
-                            instance.getDC(),
-                            instance.getHostIP(),
-                            instance.getHostName(),
-                            instance.getToken());
-                }
-
-                return instance;
-            }
-        }.call();
-    }
-
-    private PriamInstance grabDeadToken() throws Exception {
-        return new RetryableCallable<PriamInstance>() {
-            @Override
-            public PriamInstance retriableCall() throws Exception {
-                PriamInstance result = deadTokenRetriever.get();
-                if (result != null) {
-                    isReplace = true; // indicate that we are acquiring a dead instance's token.
-
-                    if (deadTokenRetriever.getReplaceIp()
-                            != null) { // The IP address of the dead instance to which
-                        // we will acquire its token
-                        replacedIp = deadTokenRetriever.getReplaceIp();
-                    }
-                }
-
-                return result;
-            }
-
-            @Override
-            public void forEachExecution() {
-                populateRacMap();
-                deadTokenRetriever.setLocMap(locMap);
-            }
-        }.call();
-    }
-
-    private PriamInstance grabPreGeneratedToken() throws Exception {
-        return new RetryableCallable<PriamInstance>() {
-            @Override
-            public PriamInstance retriableCall() throws Exception {
-                PriamInstance result = preGeneratedTokenRetriever.get();
-                if (result != null) {
-                    isTokenPregenerated = true;
-                }
-                return result;
-            }
-
-            @Override
-            public void forEachExecution() {
-                populateRacMap();
-                preGeneratedTokenRetriever.setLocMap(locMap);
-            }
-        }.call();
-    }
-
-    private PriamInstance grabNewToken() throws Exception {
-        if (!this.config.isCreateNewTokenEnable()) {
-            throw new IllegalStateException(
-                    "Node attempted to erroneously create a new token when we should be grabbing an existing token.");
-        }
-
-        return new RetryableCallable<PriamInstance>() {
-            @Override
-            public PriamInstance retriableCall() throws Exception {
-                set(100, 100);
-                newTokenRetriever.setLocMap(locMap);
-                return newTokenRetriever.get();
-            }
-
-            @Override
-            public void forEachExecution() {
-                populateRacMap();
-                newTokenRetriever.setLocMap(locMap);
-            }
-        }.call();
-    }
-
-    private Optional<PriamInstance> findInstance(
-            List<PriamInstance> instances, java.util.function.Predicate<PriamInstance> predicate) {
-        return Optional.ofNullable(instances)
-                .orElse(Collections.emptyList())
-                .stream()
-                .filter(predicate)
-                .findFirst();
-    }
-
     private void populateRacMap() {
         locMap.clear();
         List<PriamInstance> instances = factory.getAllIds(config.getAppName());
@@ -320,7 +113,7 @@ public class InstanceIdentity {
         }
     }
 
-    public List<String> getSeeds() throws UnknownHostException {
+    public List<String> getSeeds() {
         populateRacMap();
         List<String> seeds = new LinkedList<>();
         // Handle single zone deployment
