+++ priam/src/main/java/com/netflix/priam/identity/token/ITokenRetriever.java
@@ -1,0 +1,16 @@
+package com.netflix.priam.identity.token;
+
+import com.google.inject.ImplementedBy;
+import com.netflix.priam.identity.PriamInstance;
+
+/** Fetches PriamInstances and other data which is convenient at the time */
+@ImplementedBy(TokenRetriever.class)
+public interface ITokenRetriever {
+    PriamInstance get() throws Exception;
+    /** Gets the IP address of the dead instance to which we will acquire its token */
+    String getReplacedIp();
+
+    boolean isReplace();
+
+    boolean isTokenPregenerated();
+}
